Dev: refactor functions

+++ src/clean.rs
@@ -1,9 +1,12 @@
-use crate::Student;
+//use crate::Student;
 use std::fs;
 
 pub fn cleanup_student_folder() -> Result<(), Box<dyn std::error::Error>> {
     let student_output_dir = format!("./grader/");
+    if fs::read_dir(&student_output_dir)?.next().is_none() {
+        return Ok(());
+    }
     fs::remove_dir_all(&student_output_dir)?;
-    println!("已刪除學生的資料夾。");
+    println!("successfully cleanup student folder");
     Ok(())
 }
